refactor: InputArea & OutputArea 拆分组件
