+++ src/router/article-copy-tool/components/OutputFuncButtons.vue
@@ -1,0 +1,41 @@
+<script setup lang="ts">
+import {useStore} from "../../../store/useStore";
+import {useMessage} from "naive-ui";
+import useClipboard from "vue-clipboard3";
+
+const message = useMessage();
+const {toClipboard} = useClipboard();
+const store = useStore();
+
+async function copyOutputText(info: string = "复制成功") {
+  await toClipboard(store.copy.outputText);
+  message.success(info);
+}
+
+async function cutOutputText() {
+  await copyOutputText("剪切成功");
+  store.copy.outputText = "";
+}
+
+function clearOutputText() {
+  store.copy.outputText = "";
+}
+</script>
+
+<template>
+  <div class="output-func-buttons">
+    <n-space :size="30">
+      <n-button @click="copyOutputText()" type="success" :disabled="!store.copy.outputText">复制</n-button>
+      <n-button @click="cutOutputText()" type="warning" :disabled="!store.copy.outputText">剪切</n-button>
+      <n-button @click="clearOutputText()" color="#3f3f3f" :disabled="!store.copy.outputText">清空</n-button>
+    </n-space>
+  </div>
+</template>
+
+<style scoped>
+.output-func-buttons {
+  display: flex;
+  justify-content: center;
+  margin-top: 10px;
+}
+</style>
